test: 일부 테스트코드 도메인 테스트로 이동
